Fix: failing tests

+++ ci-cleanup.sh
@@ -1,4 +1,8 @@
-docker rm --force $(docker ps -aq)
+set -e
+
+if [ "$(docker ps -aq)" ]; then
+    docker rm --force $(docker ps -aq)
+fi
 
 rm -rf /tmp/logs
 mkdir -p /tmp/logs
